requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
cramjam==2.10.0
dnspython==2.7.0
filelock==3.18.0
fsspec==2025.5.0
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pillow==11.2.1
pymongo==4.13.0
python-dotenv==1.1.0
python-snappy==0.7.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.2
typing_extensions==4.13.2
urllib3==2.4.0
